Homework_1 commit - Calculator, Before and After Methods moved to separate class, var is renamed
