+++ src/test/java/com/epam/tc/hw1/BaseTest.java
@@ -1,0 +1,20 @@
+package com.epam.tc.hw1;
+
+import com.epam.tat.module4.Calculator;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
+
+public class BaseTest {
+
+    protected Calculator calculator;
+
+    @BeforeMethod
+    public void before() {
+        calculator = new Calculator();
+    }
+
+    @AfterMethod
+    public void after() {
+        calculator = null;
+    }
+}
