+++ src/test/java/com/epam/tc/hw1/DivTest.java
@@ -1,21 +1,11 @@
 package com.epam.tc.hw1;
 
-import com.epam.tat.module4.Calculator;
-
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
+
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
-public class DivTest {
-    Calculator calculator;
-
-    @BeforeMethod
-    public void before() {
-        calculator = new Calculator();
-    }
-
+public class DivTest extends BaseTest {
 
     @DataProvider
     public static Object[][] divCorrectDataDouble() {
@@ -27,14 +17,10 @@ public class DivTest {
 
     @Test(dataProvider = "divCorrectDataDouble")
     public void divTest(double a, double b, double res) {
-        var act = calculator.div(a,b);
+        double act = calculator.div(a, b);
 
         assertThat(act).as("Calculation was wrong")
                 .isEqualTo(res);
     }
 
-    @AfterMethod
-    public void after() {
-        calculator = null;
-    }
 }
