+++ src/test/java/com/epam/tc/hw1/MultTest.java
@@ -1,21 +1,11 @@
 package com.epam.tc.hw1;
 
-import com.epam.tat.module4.Calculator;
-
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
+
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
-public class MultTest {
-    Calculator calculator;
-
-    @BeforeMethod
-    public void before() {
-        calculator = new Calculator();
-    }
-
+public class MultTest extends BaseTest {
 
     @DataProvider
     public static Object[][] multCorrectDataDouble() {
@@ -28,14 +18,10 @@ public class MultTest {
 
     @Test(dataProvider = "multCorrectDataDouble")
     public void multTest(double a, double b, double res) {
-        var act = calculator.mult(a,b);
+        double act = calculator.mult(a, b);
 
         assertThat(act).as("Calculation was wrong")
                 .isEqualTo(res);
     }
 
-    @AfterMethod
-    public void after() {
-        calculator = null;
-    }
 }
