+++ src/test/java/com/epam/tc/hw1/SubTest.java
@@ -1,21 +1,11 @@
 package com.epam.tc.hw1;
 
-import com.epam.tat.module4.Calculator;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
+
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
-public class SubTest {
-
-    Calculator calculator;
-
-    @BeforeMethod
-    public void before() {
-        calculator = new Calculator();
-    }
-
+public class SubTest extends BaseTest {
 
     @DataProvider
     public static Object[][] subCorrectData() {
@@ -27,14 +17,10 @@ public class SubTest {
 
     @Test(dataProvider = "subCorrectData")
     public void subTest(double a, double b, double res) {
-        var act = calculator.sub(a,b);
+        double act = calculator.sub(a, b);
 
         assertThat(act).as("Calculation was wrong")
                 .isEqualTo(res);
     }
 
-    @AfterMethod
-    public void after() {
-        calculator = null;
-    }
 }
